audit-consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.3.0
python-dotenv==1.0.0